Satt sammen Map og customMapViewer klassene CustomMapViewer kan nå si ifra når en MouseEvent skjer.
